Refactor: refactor slider component and place detail page 슬라이더 컴포넌트 및 장소 세부정보 페이지 리팩토리
