+++ frontend/pages/mypage/wishlist/[placeid]/index.tsx
@@ -1,4 +1,4 @@
-import PlaceDetail from "../../../../src/components/units/mypage/placedetail/PlaceDetail.container";
+import PlaceDetail from "../../../../src/components/commons/placedetail/PlaceDetail.container";
 
 export default function WishListDetail(): JSX.Element {
     return (
