+++ frontend/src/components/commons/placedetail/PlaceDetail.container.tsx
