+++ frontend/src/components/commons/placedetail/PlaceDetail.presenter.tsx
