+++ frontend/src/components/commons/placedetail/PlaceDetail.styles.tsx
