+++ frontend/src/components/commons/ratedplace/RatedPlaceSlider.container.tsx
