+++ frontend/src/components/commons/ratedplace/RatedPlaceSlider.types.ts
