+++ frontend/src/components/commons/wishlist/WishListSlider.styles.tsx
